requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.3.0
xgboost>=1.7.6
lightgbm>=4.1.0
nba_api>=1.4.1
requests>=2.26.0
python-dotenv>=0.19.0
joblib>=1.3.0 
beautifulsoup4
selenium
webdriver-manager
